+++ code/src/ft2/ftgrays.c
@@ -1063,7 +1063,7 @@ void  quick_sort( PCell cells,
 
 	for (;; )
 	{
-		int len = limit - base;
+		int len = (int)( limit - base );
 		PCell i, j, pivot;
