+++ code/src/jpeg-6/jmemnobs.c
@@ -29,7 +29,7 @@
 
 GLOBAL void *
 jpeg_get_small( j_common_ptr cinfo, size_t sizeofobject ) {
-	return Z_Malloc( sizeofobject );
+	return Z_Malloc( (int)sizeofobject );
 }
 
 GLOBAL void
@@ -47,7 +47,7 @@ jpeg_free_small( j_common_ptr cinfo, void * object, size_t sizeofobject ) {
 
 GLOBAL void FAR *
 jpeg_get_large( j_common_ptr cinfo, size_t sizeofobject ) {
-	return Z_Malloc( sizeofobject );
+	return Z_Malloc( (int)sizeofobject );
 }
 
 GLOBAL void
